Protect memoization ivar setup with a global mutex
`setup_memoization_method_name` can be preempted by another thread at
anytime, and then continued later. This means the memoized value can
be set to `nil` after it was set to a value by another thread.

Protect the ivar setup with a global mutex to prevent this race
condition.

Fixes #4